+++ app/src/main/java/ch/epfl/sdp/appart/database/MockDatabaseService.java
@@ -31,7 +31,7 @@ public class MockDatabaseService implements DatabaseService {
     private final List<String> images = new ArrayList<>();
 
     public MockDatabaseService() {
-        cards.add(new Card("1111", "unknown", "unknown", "Lausanne", 1000, "file:///android_asset/apart_fake_image_1.jpeg"));
+        cards.add(new Card("1111", "unknown", "5555", "Lausanne", 1000, "file:///android_asset/apart_fake_image_1.jpeg"));
         cards.add(new Card("2222", "unknown", "unknown", "Lausanne", 1000, "file:///android_asset/apart_fake_image_1.jpeg"));
         cards.add(new Card("3333", "unknown", "unknown", "Lausanne", 1000, "file:///android_asset/apart_fake_image_1.jpeg"));
         cards.add(new Card("4444", "unknown", "unknown", "Lausanne", 1000, "file:///android_asset/apart_fake_image_1.jpeg"));
